refactor: move game canvas to client component
